ScriptPy/requirement.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cssselect==1.1.0
dnspython==2.1.0
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.8.2
lxml==4.6.4
numpy==1.21.4
pandas==1.3.4
parse==1.19.0
pyee==8.2.2
pymongo==3.12.1
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
urllib3==1.26.7
w3lib==1.22.0
websockets==9.1
zipp==3.6.0
